% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calculate.R
\name{icc_sample_size_width}
\alias{icc_sample_size_width}
\title{Sample Size for ICC based on Confidence Interval Width}
\usage{
icc_sample_size_width(
  rho,
  omega,
  k = 3,
  same_raters,
  rater_effect = NULL,
  rating_type,
  agreement_type = NULL,
  alpha = 0.05,
  assurance = 0.8,
  verbose = TRUE
)
}
\arguments{
\item{rho}{Anticipated ICC value.}

\item{omega}{Desired half-width.}

\item{k}{Number of observations. Default 3.}

\item{same_raters}{Logical.}

\item{rater_effect}{"random" or "fixed".}

\item{rating_type}{"single" or "average".}

\item{agreement_type}{"absolute" or "consistency".}

\item{alpha}{Significance level. Default 0.05.}

\item{assurance}{Assurance probability. Default 0.8.}

\item{verbose}{Print messages. Default TRUE.}
}
\value{
Required sample size.
}
\description{
Sample Size for ICC based on Confidence Interval Width
}
\examples{
# Note: It is recommended to use the unified interface icc_sample_size()
icc_sample_size_width(rho = 0.7, omega = 0.1, k = 3, same_raters = FALSE, rating_type = "average")
}
