% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{icc_shiny_app}
\alias{icc_shiny_app}
\title{Build the ICCDesign Shiny Application}
\usage{
icc_shiny_app()
}
\value{
A Shiny application object.
}
\description{
Build the ICCDesign Shiny Application
}
\keyword{internal}
