% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc_analysis.R
\name{print.icc_result}
\alias{print.icc_result}
\title{打印 ICC 结果对象}
\usage{
\method{print}{icc_result}(x, ...)
}
\arguments{
\item{x}{\code{icc_result} 类对象}

\item{...}{其他参数（未使用）}
}
\description{
打印 ICC 结果对象
}
