% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size_calculate.R
\name{icc_sample_size}
\alias{icc_sample_size}
\title{Unified ICC Sample Size & Power Interface}
\usage{
icc_sample_size(method = c("lower", "width", "power"), ...)
}
\arguments{
\item{method}{"lower", "width", "power".}

\item{...}{Arguments passed to underlying functions.}
}
\value{
Sample size or power.
}
\description{
Unified ICC Sample Size & Power Interface
}
